# docs/api-spec.yaml
openapi: "3.0.3"
info:
  title: Crawler.io API
  description: Unified Social Media Data API — one integration, all platforms.
  version: "1.0.0"
  contact:
    email: api@crawler.io
    url: https://docs.crawler.io
servers:
  - url: https://api.crawler.io/v1
    description: Production
  - url: https://sandbox.crawler.io/v1
    description: Sandbox (mock data)
security:
  - BearerAuth: []
...
paths:
...
  /creators/lookup:
    get:
      tags: [Creators]
      summary: Lookup creator by platform username
      parameters:
        - name: platform
          in: query
          required: true
          schema:
            type: string
            enum: [instagram, tiktok, youtube]
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Creator found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
              example:
                id: "cr_a1b2c3d4e5"
                platform: "instagram"
                platform_id: "12345678"
                username: "johndoe"
                display_name: "John Doe"
                bio: "Travel photographer | NYC"
                is_verified: true
                follower_count: 1250000
                following_count: 843
                post_count: 2341
                engagement_rate: 3.42
                data_freshness: "live"
                last_updated_at: "2026-03-31T14:22:10Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /creators/{creator_id}:
    get:
      tags: [Creators]
      summary: Get creator by internal ID
      parameters:
        - name: creator_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^cr_[a-z0-9]+$"
      responses:
        "200":
          description: Creator details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "404":
          $ref: "#/components/responses/NotFound"
  /creators/search:
    get:
      tags: [Creators]
      summary: Search creators
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Free-text search across username, display name, bio
        - name: platform
          in: query
          schema:
            type: string
            enum: [instagram, tiktok, youtube]
        - name: min_followers
          in: query
          schema:
            type: integer
        - name: max_followers
          in: query
          schema:
            type: integer
        - name: verified
          in: query
          schema:
            type: boolean
        - name: category
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCreators"
...
  /creators/bulk-lookup:
    post:
      tags: [Creators]
      summary: Bulk lookup creators (up to 100)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lookups:
                  type: array
                  maxItems: 100
                  items:
                    type: object
                    required: [platform, username]
                    properties:
                      platform:
                        type: string
                        enum: [instagram, tiktok, youtube]
                      username:
                        type: string
      responses:
        "200":
          description: Bulk results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        platform:
                          type: string
                        username:
                          type: string
                        creator:
                          $ref: "#/components/schemas/Creator"
                        error:
                          type: string

  # ─── POSTS ──────────────────────────────────
...
  /tracking/creators/{tracking_id}:
    delete:
      tags: [Tracking]
      summary: Stop tracking a creator
      parameters:
        - name: tracking_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tracking stopped

  # ─── WEBHOOKS ───────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Pass your key as a Bearer token.
        Production keys: sk_live_*
        Sandbox keys: sk_test_*
  schemas:
    Creator:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
        platform_id:
          type: string
        username:
          type: string
        display_name:
          type: string
        bio:
          type: string
        profile_image_url:
          type: string
        is_verified:
          type: boolean
        follower_count:
          type: integer
        following_count:
          type: integer
        post_count:
          type: integer
        engagement_rate:
          type: number
        categories:
          type: array
          items:
            type: string
        external_urls:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
        first_seen_at:
          type: string
          format: date-time
        last_updated_at:
          type: string
          format: date-time
        data_freshness:
          type: string
          enum: [live, recent, stale]
...
    Pagination:
      type: object
      properties:
        has_more:
          type: boolean
        next_cursor:
          type: string
        total_count:
          type: integer
    PaginatedCreators:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Creator"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            status:
              type: integer
            retry_after:
              type: integer
        request_id:
          type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "not_found"
              message: "Creator not found"
              status: 404
            request_id: "req_abc123"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "rate_limit_exceeded"
              message: "You have exceeded your rate limit"
              status: 429
              retry_after: 32
            request_id: "req_abc123"